having issues loading the js module into the html for some reason - busy investigating, have attempted all documented styles and still not working as expected.

+++ src/scripts/main.js
@@ -1,18 +1,34 @@
-import DataService from "./modules/data_service";
-import SessionStore from "./modules/session_store";
+import DataService from "./modules/data_service.js";
+import SessionStore from "./modules/session_store.js";
 
 const dataService = new DataService();
 const sessionStore = new SessionStore();
 
-const createUser = (name, email) => {
+const createUser = () => {
   const time = new Date().getTime();
-  const user = dataService.create(name, email, time);
-  sessionStore.init(user);
+  // const user = dataService.create(name, email, time);
+  sessionStore.init(name, email, time);
 }
-
 const removeUser = (email) => {
+  if(!email) {
+    throw new Error("Email is required");
+  }
   dataService.delete(email);
   sessionStore.destroy();
 };
 
-document.addEventListener("beforeunload", removeUser(sessionStore.email));
+const submitForm = (e) => {
+  e.preventDefault();
+  const formData = new FormData(e.target);
+  const name = formData.get("name");
+  const email = formData.get("email");
+  console.log(name, email);
+  createUser(name, email);
+};
+
+window.addEventListener("beforeunload", (e) => {
+  console.log(e);
+  removeUser(sessionStore.email)
+});
+
+export default {submitForm, removeUser, createUser};
